+++ authentication-service/src/main/java/com/ultimatesoftware/banking/authentication/service/AuthenticationServiceApplication.java
@@ -1,0 +1,19 @@
+package com.ultimatesoftware.banking.authentication.service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@SpringBootApplication
+public class AuthenticationService {
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(AuthenticationService.class, args);
+    }
+}
