+++ authentication-service/src/main/java/com/ultimatesoftware/banking/authentication/service/security/JWTSuccessHandler.java
@@ -1,0 +1,17 @@
+package com.ultimatesoftware.banking.authentication.service.security;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+public class JWTSuccessHandler implements AuthenticationSuccessHandler{
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, Authentication authentication) throws IOException, ServletException {
+        System.out.println("Successfully Authentication");
+    }
+}
