+++ authentication-service/src/test/java/com/ultimatesoftware/banking/authentication/service/AuthenticationServiceApplicationTest.java
@@ -1,0 +1,4 @@
+package com.ultimatesoftware.banking.authentication.service;
+
+public class AuthenticationServiceTest {
+}
