--- authorization-service/build.gradle
@@ -1,35 +1,0 @@
-group 'com.ultimatesoftware.banking'
-version = '4.6'
-
-apply plugin: 'java'
-apply plugin: 'idea'
-apply plugin: 'maven'
-apply plugin: 'application'
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
-
-mainClassName = 'com.ultimatesoftware.banking.authorization.service.AuthorizationService'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-checkstyle {
-    sourceSets = []
-}
-
-dependencies {
-    compile 'io.jsonwebtoken:jjwt:0.9.0'
-    compile("org.springframework.boot:spring-boot-starter-security")
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '1.5.14.RELEASE'
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-data-mongodb', version: '2.0.3.RELEASE'
-
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.0.3.RELEASE'
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = "${version}"
-}
