--- authorization-service/src/main/java/com/ultimatesoftware/banking/authorization/service/AuthorizationService.java
@@ -1,19 +1,0 @@
-package com.ultimatesoftware.banking.authorization.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@SpringBootApplication
-public class AuthorizationService {
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuthorizationService.class, args);
-    }
-}
