--- authorization-service/src/main/java/com/ultimatesoftware/banking/authorization/service/config/MongoDBConfig.java
@@ -1,13 +1,0 @@
-package com.ultimatesoftware.banking.authorization.service.config;
-
-import com.ultimatesoftware.banking.authorization.service.services.UserRepository;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@EnableMongoRepositories(basePackageClasses = UserRepository.class)
-@Configuration
-public class MongoDBConfig {
-
-
-}
-
