--- authorization-service/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-# Mongo config
-spring:
-  data.mongodb:
-    host: ${MONGO_HOST:localhost}
-    port: 27017
-    database: authorization
-    repositories:
-      enabled: true
-
-  application:
-    name: authorization-service
-  resources:
-    chain:
-      enabled: true
-
-server:
-  port: 8080
-
-#logging
-logging:
-  level:
-    org.springframework.web: ERROR
-    com.mkyong: DEBUG
-
